+++ src/DGtal/io/Display2DFactory.h
@@ -59,7 +59,7 @@
 #include "DGtal/topology/KhalimskySpaceND.h"
 #include "DGtal/topology/Object.h"
 #include "DGtal/kernel/PointVector.h"
-#include "DGtal/geometry/curves/tools/Preimage2D.h"
+#include "DGtal/geometry/tools/Preimage2D.h"
 #include "DGtal/shapes/fromPoints/StraightLineFrom2Points.h"
 
 #include "DGtal/io/boards/Board2D.h"
